+++ post-install.d/git.sh
@@ -21,11 +21,11 @@ which nano  &>/dev/null && GIT_EDITORS+=("nano")
 which gedit &>/dev/null && GIT_EDITORS+=("gedit")
 select GIT_EDITOR in ${GIT_EDITORS[@]}; do
 case $GIT_EDITOR in
-	vscode) git config --global core.editor "code --wait"                                ;;
-	vim)    git config --global core.editor "vim -n -c 'no undofile' -c 'set nobackup'"  ;;
-	nano)   git config --global core.editor "nano"                                       ;;
-	gedit)  git config --global core.editor "gedit -s"                                   ;;
-	*) printf "Option %s not recognized.\n" $GIT_EDITOR; continue                        ;;
+	vscode) git config --global core.editor "code --wait"                                  ;;
+	vim)    git config --global core.editor "vim -n -c 'set noundofile' -c 'set nobackup'" ;;
+	nano)   git config --global core.editor "nano"                                         ;;
+	gedit)  git config --global core.editor "gedit -s"                                     ;;
+	*) printf "Option %s not recognized.\n" $GIT_EDITOR; continue                          ;;
 esac; break; done
 unset GIT_EDITOR GIT_EDITORS